move string related methods from Util to more specific StringUtils class
